Reorganize Collections to be part of Administrative tab in Work and Batch edit.  Also cleaned up Adminstrative tab form layout for consistency between Work and Batch edit modes.
