+++ assets/js/components/BatchEdit/About/About.jsx
@@ -31,7 +31,7 @@ const BatchEditAbout = () => {
   const [batchReplaces, setBatchReplaces] = useState({
     descriptiveMetadata: {},
   });
-  const [batchCollection, setBatchCollection] = useState({});
+
   const [batchPublish, setBatchPublish] = useState({
     publish: false,
     unpublish: false,
@@ -115,10 +115,6 @@ const BatchEditAbout = () => {
         published: { ...batchPublish },
       }),
     });
-
-    Object.keys(currentFormValues["collection"]).length > 0
-      ? setBatchCollection(JSON.parse(currentFormValues["collection"]))
-      : setBatchCollection({});
 
     setIsConfirmModalOpen(true);
   };
@@ -205,7 +201,6 @@ const BatchEditAbout = () => {
           batchAdds={batchAdds}
           batchDeletes={batchDeletes}
           batchReplaces={batchReplaces}
-          batchCollection={batchCollection}
           filteredQuery={JSON.stringify(batchState.filteredQuery)}
           handleClose={onCloseModal}
           handleFormReset={handleFormReset}
