+++ assets/js/components/BatchEdit/About/CoreMetadata.jsx
@@ -1,23 +1,14 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { useQuery } from "@apollo/client";
 import UITagNotYetSupported from "../../UI/TagNotYetSupported";
 import UIInput from "../../UI/Form/Input";
 import UIFormField from "../../UI/Form/Field";
 import UIFormBatchFieldArray from "../../UI/Form/BatchFieldArray";
-import { GET_COLLECTIONS } from "@js/components/Collection/collection.gql";
 import { useFormContext } from "react-hook-form";
 import { useCodeLists } from "@js/context/code-list-context";
 
 const BatchEditAboutCoreMetadata = ({ ...restProps }) => {
   const codeLists = useCodeLists();
-
-  const {
-    loading: collectionLoading,
-    error: collectionError,
-    data: collectionData,
-  } = useQuery(GET_COLLECTIONS);
-
   const context = useFormContext();
   const register = context.register;
 
@@ -40,21 +31,21 @@ const BatchEditAboutCoreMetadata = ({ ...restProps }) => {
       </div>
 
       <div className="column is-half">
+        {/* Alternate Title */}
+        <UIFormBatchFieldArray
+          name="alternateTitle"
+          data-testid="alternate-title"
+          label="Alternate Title"
+        />
+      </div>
+
+      <div className="column is-half">
         {/* Description */}
         <UIFormBatchFieldArray
           name="description"
           label="Description"
           data-testid="description"
           isTextarea={true}
-        />
-      </div>
-
-      <div className="column is-half">
-        {/* Alternate Title */}
-        <UIFormBatchFieldArray
-          name="alternateTitle"
-          data-testid="alternate-title"
-          label="Alternate Title"
         />
       </div>
 
@@ -93,25 +84,6 @@ const BatchEditAboutCoreMetadata = ({ ...restProps }) => {
           />
           <UITagNotYetSupported label="Display not yet supported" />
           <UITagNotYetSupported label="Update not yet supported" />
-        </UIFormField>
-      </div>
-
-      <div className="column is-half">
-        <UIFormField label="Collection">
-          <div className="select">
-            <select name="collection" ref={register()} data-testid="collection">
-              <option value="">-- Select --</option>
-              {collectionData &&
-                collectionData.collections.map((collection) => (
-                  <option
-                    key={collection.id}
-                    value={JSON.stringify(collection)}
-                  >
-                    {collection.title}
-                  </option>
-                ))}
-            </select>
-          </div>
         </UIFormField>
       </div>
     </div>
