+++ assets/js/components/BatchEdit/Administrative/Administrative.test.jsx
@@ -5,6 +5,7 @@ import { renderWithRouterApollo } from "../../../services/testing-helpers";
 import { allCodeListMocks } from "@js/components/Work/controlledVocabulary.gql.mock";
 import { BatchProvider } from "@js/context/batch-edit-context";
 import { CodeListProvider } from "@js/context/code-list-context";
+import { getCollectionsMock } from "@js/components/Collection/collection.gql.mock";
 
 const items = ["ABC123", "ZYC889"];
 
@@ -17,7 +18,7 @@ describe("BatchEditAdministrative component", () => {
         </CodeListProvider>
       </BatchProvider>,
       {
-        mocks: allCodeListMocks,
+        mocks: [...allCodeListMocks, getCollectionsMock],
       }
     );
   }
@@ -32,23 +33,28 @@ describe("BatchEditAdministrative component", () => {
   it("renders the sticky header", async () => {
     const { getByTestId } = setupTest();
     await waitFor(() => {
-      expect(
-        getByTestId("batch-edit-administrative-sticky-header")
-      ).toBeInTheDocument();
+      expect(getByTestId("batch-edit-administrative-sticky-header"));
+    });
+  });
+
+  it("renders the Batch Collection component", async () => {
+    const { getByTestId } = setupTest();
+    await waitFor(() => {
+      expect(getByTestId("batch-collection-wrapper"));
     });
   });
 
   it("renders project metadata component", async () => {
     const { getByTestId } = setupTest();
     await waitFor(() => {
-      expect(getByTestId("project-metadata")).toBeInTheDocument();
+      expect(getByTestId("project-metadata"));
     });
   });
 
-  it("renders project status metadata component", async () => {
+  it("renders general component", async () => {
     const { getByTestId } = setupTest();
     await waitFor(() => {
-      expect(getByTestId("project-status-metadata")).toBeInTheDocument();
+      expect(getByTestId("project-status-metadata"));
     });
   });
 });
