+++ assets/js/components/BatchEdit/Administrative/General.test.jsx
@@ -4,16 +4,14 @@ import {
   renderWithRouterApollo,
   withReactHookForm,
 } from "../../../services/testing-helpers";
-import BatchEditAdministrativeProjectStatusMetadata from "./ProjectStatusMetadata";
+import BatchEditAdministrativeGeneral from "./General";
 import { BatchProvider } from "../../../context/batch-edit-context";
 import { CodeListProvider } from "@js/context/code-list-context";
 import { allCodeListMocks } from "@js/components/Work/controlledVocabulary.gql.mock";
 
-describe("BatchEditAdministrativeProjectStatusMetadata component", () => {
+describe("BatchEditAdministrativeGeneral component", () => {
   beforeEach(() => {
-    const Wrapped = withReactHookForm(
-      BatchEditAdministrativeProjectStatusMetadata
-    );
+    const Wrapped = withReactHookForm(BatchEditAdministrativeGeneral);
     return renderWithRouterApollo(
       <BatchProvider value={null}>
         <CodeListProvider>
@@ -33,7 +31,6 @@ describe("BatchEditAdministrativeProjectStatusMetadata component", () => {
       "preservationLevel",
       "status",
       "visibility",
-      "projectCycle",
       "libraryUnit",
     ];
     for (let item of itemTestIds) {
